# compile integration tests as a single binary