--- a/cli/tests/test_checkout.rs
+++ b/cli/tests/test_checkout.rs
@@ -16,8 +16,6 @@ use std::path::Path;
 
 use crate::common::TestEnvironment;
 
-pub mod common;
-
 #[test]
 fn test_checkout() {
     let test_env = TestEnvironment::default();