--- a/cli/tests/test_describe_command.rs
+++ b/cli/tests/test_describe_command.rs
@@ -14,8 +14,6 @@
 
 use crate::common::{get_stderr_string, TestEnvironment};
 
-pub mod common;
-
 #[test]
 fn test_describe() {
     let mut test_env = TestEnvironment::default();