--- a/cli/tests/test_interdiff_command.rs
+++ b/cli/tests/test_interdiff_command.rs
@@ -12,9 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use common::TestEnvironment;
-
-pub mod common;
+use crate::common::TestEnvironment;
 
 #[test]
 fn test_interdiff_basic() {