--- a/cli/tests/test_move_command.rs
+++ b/cli/tests/test_move_command.rs
@@ -16,8 +16,6 @@ use std::path::Path;
 
 use crate::common::TestEnvironment;
 
-pub mod common;
-
 #[test]
 fn test_move() {
     let test_env = TestEnvironment::default();