--- a/cli/tests/test_revset_output.rs
+++ b/cli/tests/test_revset_output.rs
@@ -12,9 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use common::TestEnvironment;
-
-pub mod common;
+use crate::common::TestEnvironment;
 
 #[test]
 fn test_syntax_error() {