--- a/cli/tests/test_show_command.rs
+++ b/cli/tests/test_show_command.rs
@@ -12,11 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use common::TestEnvironment;
 use itertools::Itertools;
 use regex::Regex;
 
-pub mod common;
+use crate::common::TestEnvironment;
 
 #[test]
 fn test_show() {