--- a/cli/tests/test_templater.rs
+++ b/cli/tests/test_templater.rs
@@ -16,8 +16,6 @@ use std::path::Path;
 
 use crate::common::TestEnvironment;
 
-pub mod common;
-
 #[test]
 fn test_templater_parse_error() {
     let test_env = TestEnvironment::default();