--- a/cli/tests/test_untrack_command.rs
+++ b/cli/tests/test_untrack_command.rs
@@ -16,8 +16,6 @@ use std::path::PathBuf;
 
 use crate::common::TestEnvironment;
 
-pub mod common;
-
 #[test]
 fn test_untrack() {
     let test_env = TestEnvironment::default();