--- a/lib/Cargo.toml
+++ b/lib/Cargo.toml
@@ -1,6 +1,7 @@
 [package]
 name = "jj-lib"
 description = "Library for Jujutsu - an experimental version control system"
+autotests = false
 
 version = { workspace = true }
 edition = { workspace = true }
@@ -10,6 +11,9 @@ homepage = { workspace = true }
 repository = { workspace = true }
 documentation = { workspace = true }
 readme = { workspace = true }
+
+[[test]]
+name = "runner"
 
 [[bench]]
 name = "diff_bench"