@ilyagr 2024-02-03
The benchmark in the commit message is `touch lib/src/lib.rs ; cargo t --test runner --no-run`, and I'm assuming `hyperfine` throws out the first run like a good benchmarking tool should. So, it's all about incremental compilation, I think.