@ilyagr 2024-02-03
I also unscientifically tried `touch lib/src/lib.rs ; cargo insta test --test-runner nextest -- test_abandon`. It feels fast with this PR, though which linker you use might again make a difference.

I found another minor downside: the fish shell has completions for `cargo test --test <TAB>`, but not for `cargo test -- <TAB>` (or they don't work with this PR). In this way, doing a subset of tests is less convenient for me with this PR (though I think the speedup is worth it for me).