@arxanas 2024-02-03
Regarding `touch`, @jyn514 would probably know better than me, but in my blog post I actually modified the file in the worry that `cargo` might do something content-addressed as part of compilation and it wouldn't represent a "real" incremental change.