@jyn514 2024-02-03
> I found another minor downside: the fish shell has completions for `cargo test --test <TAB>`, but not for `cargo test -- <TAB>` (or they don't work with this PR). In this way, doing a subset of tests is less convenient for me with this PR (though I think the speedup is worth it for me).

yeah, this is unfortunate. i don't know a way to fix this upstream in fish; it would have to parse the `mod` declarations (which is not *super* hard but requires it to be aware of cargo's directory structure instead of just running `cargo metadata`).