@jyn514 2024-02-03
> Can you compare the resulting runtimes for running the whole test suite

it takes about 13.5 seconds for me to run the tests after they're compiled, both before and after this PR. so the total time is going from ~45 to ~18 seconds (and it's better than 18 in practice because it's rare for the very last test run to be the only one that fails).